requirements.txt
info-nce-pytorch==0.1.4
numpy==1.26.3
pandas==2.2.0
pytorch-metric-learning==2.4.1
scikit-learn==1.4.0
torch==2.2.0
torch_geometric==2.4.0
tqdm==4.66.1
transformers==4.37.2
